// sourcecode/BTrak.Packages/snova-app-builder-creation-components/app-builder-module-components/app-builder-components/src/lib/builder/models/file-input.model.ts
export class FileInputModel {
    fileId: string;
    fileName: string;
    fileSize: number;
    filePath: string;
    fileExtension: string;
    isArchived: boolean;
    folderId: string;
    storeId: string;
    referenceId: string;
    referenceTypeName: string;
    fileType: number;
    createdDateTime: Date;
    createdByUserId: string;
    referenceTypeId: string;
    timeStamp: any;
    isDefault: boolean;
    userStoryId: string;
    isToBeReviewed: boolean;
    reviewedDateTime: Date;
    reviewedByUser: string;
    isTreeView: boolean;
}